Fix multiple experiment failures
1. FlowMatchingBridge dtype fix:
   - Added dtype to torch.tensor() in integrate_flow
   - Fixed velocity_net dtype matching in compute_velocity

2. Data batching fix:
   - Added custom_collate_fn for variable-length nested data (ARC choices)
   - Handles HuggingFace datasets with nested dicts properly

3. ARC dataset 5-choice support:
   - Updated labels to include "E" (some ARC questions have 5 choices)
   - Updated label_map in run_baselines.py
   - Updated num_classes to 5

4. Prompt tuning AttributeError fix:
   - Handle both tensor and list label fields (tolist() only if tensor)

+++ telepathy/cross_model_experiments.py
@@ -1143,11 +1143,11 @@ class FlowMatchingBridge(nn.Module):
             t = t.expand(B, 1)
         t_expanded = t.unsqueeze(-1).expand(B, K, 1)
 
-        # Concatenate position and time
-        xt = torch.cat([x, t_expanded], dim=-1)  # [B, K, D+1]
+        # Concatenate position and time (ensure dtype matches velocity_net weights)
+        xt = torch.cat([x, t_expanded.to(x.dtype)], dim=-1)  # [B, K, D+1]
 
-        # Compute velocity
-        v = self.velocity_net(xt)
+        # Compute velocity (match dtype to weights)
+        v = self.velocity_net(xt.to(self.velocity_net[0].weight.dtype))
         return v
 
     def integrate_flow(self, x0: torch.Tensor, num_steps: int = None) -> torch.Tensor:
@@ -1166,7 +1166,7 @@ class FlowMatchingBridge(nn.Module):
         dt = 1.0 / num_steps
 
         for i in range(num_steps):
-            t = torch.tensor(i * dt, device=x.device)
+            t = torch.tensor(i * dt, device=x.device, dtype=x.dtype)
             v = self.compute_velocity(x, t)
             x = x + dt * v
